Iologindata transform integer change

+++ iologindata.cpp
@@ -469,7 +469,7 @@ bool IOLoginData::loadPlayer(Player* player, DBResult_ptr result)
 	player->currentOutfit = player->defaultOutfit;
 	player->direction = static_cast<Direction> (result->getNumber<uint16_t>("direction"));
 	player->setSpecialLootRate(result->getNumber<uint16_t>("special_rate_loot"));
-	player->setTransform(result->getNumber<uint16_t>("transform"));
+	player->setTransform(result->getNumber<uint8_t>("transform"));
 
 	if (g_game.getWorldType() != WORLD_TYPE_PVP_ENFORCED) {
 		const time_t skullSeconds = result->getNumber<time_t>("skulltime") - time(nullptr);
